Use int16 for eeprom "address"

+++ Marlin/src/HAL/ESP32/eeprom.cpp
@@ -36,7 +36,7 @@ size_t PersistentStore::capacity()    { return MARLIN_EEPROM_SIZE; }
 bool PersistentStore::access_start()  { return EEPROM.begin(MARLIN_EEPROM_SIZE); }
 bool PersistentStore::access_finish() { EEPROM.end(); return true; }
 
-bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, uint16_t *crc) {
+bool PersistentStore::write_data(uint16_t &pos, const uint8_t *value, size_t size, uint16_t *crc) {
   for (size_t i = 0; i < size; i++) {
     EEPROM.write(pos++, value[i]);
     crc16(crc, &value[i], 1);
@@ -44,7 +44,7 @@ bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, ui
   return false;
 }
 
-bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
+bool PersistentStore::read_data(uint16_t &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
   for (size_t i = 0; i < size; i++) {
     uint8_t c = EEPROM.read(pos++);
     if (writing) value[i] = c;
